server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
httpx==0.25.2
pytest==7.4.3
black==23.12.0
isort==5.13.2
flake8==6.1.0
